commented on Input

+++ core/src/ghozti/racoonadventure/structure/Input.java
@@ -5,7 +5,15 @@ import com.badlogic.gdx.Input.Keys;
 
 public class Input {
 
-    public static float[] HandleInput(float x, float y, int speed, float deltaTime){
+    /**
+     * Handle input will:
+     * take in 4 floats:
+     * x,y,speed and delta time
+     * speed represents the amount of px to move
+     * this method will return the updates x and y value based on the key inputed
+     */
+
+    public static float[] HandleInput(float x, float y, float speed, float deltaTime){
         if (Gdx.input.isKeyPressed(Keys.W)){//move forward
             y += speed/2;
         }
@@ -18,7 +26,6 @@ public class Input {
         if (Gdx.input.isKeyPressed(Keys.D)){//move to the right
             x += speed;
         }
-
         return new float[]{x,y};
     }
 }
